+++ src/app/core/models/comment.ts
@@ -1,61 +1,3 @@
-export interface IEventComment {
-  id: string;
-  eventId: string;
-  commentatorDisplayedName: string;
-  commentatorDisplayedImage: string;
-  userUid: string;
-  userName: string;
-  userImage: string;
-  message: string;
-  timestamp: string;
-}
-
-
-export class EventComment implements IEventComment {
-
-  constructor(
-    public id: string,
-    public eventId: string,
-    public commentatorDisplayedName: string,
-    public commentatorDisplayedImage: string,
-    public userUid: string,
-    public userName: string,
-    public userImage: string,
-    public message: string,
-    public timestamp: string
-  ) {
-  }
-}
-
-export interface INoticeComment {
-  id: string;
-  noticeId: string;
-  commentatorDisplayedName: string;
-  commentatorDisplayedImage: string;
-  userUid: string;
-  userName: string;
-  userImage: string;
-  message: string;
-  timestamp: string;
-}
-
-export class NoticeComment implements INoticeComment {
-
-  constructor(
-    public id: string,
-    public noticeId: string,
-    public commentatorDisplayedName: string,
-    public commentatorDisplayedImage: string,
-    public userUid: string,
-    public userName: string,
-    public userImage: string,
-    public message: string,
-    public timestamp: string
-  ) {
-  }
-}
-
-
 export interface IComment {
   id: string;
   itemId: string;
@@ -66,6 +8,7 @@ export interface IComment {
   userImage: string;
   message: string;
   timestamp: string;
+  commentType: CommentType;
 }
 
 
@@ -80,7 +23,22 @@ export class Comment implements IComment {
     public userName: string,
     public userImage: string,
     public message: string,
-    public timestamp: string
+    public timestamp: string,
+    public commentType: CommentType
   ) {
   }
 }
+
+
+// eslint-disable-next-line no-shadow
+export enum CommentType {
+  Event = 'EVENT',
+  Notice = 'NOTICE',
+}
+
+const COMMENT_TYPES: CommentType[] = [
+  CommentType.Event,
+  CommentType.Notice,
+];
+
+export { COMMENT_TYPES };